Move machine page content to home page and update navigation
